Keeping Track of Drink Button Pushes
Keep track of the drink button pushes for the user. They are unable to
press "drink" if all the players are not ranked on the team. Before they
hit drink they can rearrange the rankings if needed.

Added more team object properties to keep track of information on the
team.

More comments in the handlebar templates