@font-face working
Figured out the path issue where if the path to the font files were
incorrect meteor would still return a 200 OK message to the browser.

Bootstrap-less3 assigns the path for the font files to "../fonts/" which
should be changed to "/fonts/". You then move the /fonts folder into the
/public folder in the root of the Meteor project. Then meteor makes
those files available in the root. Then it'll work correctly.